Add useGetSelectedNode hook
